+++ BookCataloque/BookCataloque.SSDT/dbo/Tables/Books.sql
@@ -2,7 +2,7 @@
     [BookID]          INT            IDENTITY (1, 1) NOT NULL,
     [Title]           NVARCHAR (100) NOT NULL,
     [PublicationDate] DATE           NOT NULL,
-    [Rating]          TINYINT        NULL,
+    [Rating]          NUMERIC(4, 2)        NULL,
     [Pages] SMALLINT NOT NULL, 
     CONSTRAINT [PK_Books] PRIMARY KEY CLUSTERED ([BookID] ASC)
 );
